+++ legacy/pre-erc20-Jun2025/marketplace/ui/src/components/ETHST/EthstProductDetails.js
@@ -292,12 +292,12 @@ const ProductDetails = ({ user, users }) => {
   let fileValues = [];
   let fileNames = [];
 
-  if (details && Array.isArray(details['BlockApps-Mercata-Asset-fileNames'])) {
-    fileNames = details['BlockApps-Mercata-Asset-fileNames'];
+  if (details && Array.isArray(details['BlockApps-Asset-fileNames'])) {
+    fileNames = details['BlockApps-Asset-fileNames'];
   }
 
-  if (details && Array.isArray(details['BlockApps-Mercata-Asset-files'])) {
-    fileValues = details['BlockApps-Mercata-Asset-files'].map((file, index) => {
+  if (details && Array.isArray(details['BlockApps-Asset-files'])) {
+    fileValues = details['BlockApps-Asset-files'].map((file, index) => {
       let name = fileNames[index]?.value
         ? fileNames[index]?.value
         : `Information-${index + 1}.pdf`;
@@ -556,7 +556,7 @@ const ProductDetails = ({ user, users }) => {
           <EthstSteps />
           <div className="flex w-full flex-col md:leading-12 px-4 sm:px-8 md:px-0 items-center md:w-[750px] md:w-[835px] xl:w-[858px]  md:mx-auto mt-12">
             <div className="flex md:justify-center gap-[15px] md:gap-6 flex-col md:flex-row items-center">
-              {details['BlockApps-Mercata-Asset-images'].length > 0 ? (
+              {details['BlockApps-Asset-images'].length > 0 ? (
                 <Swiper
                   spaceBetween={30}
                   effect={'fade'}
@@ -568,8 +568,8 @@ const ProductDetails = ({ user, users }) => {
                   modules={[Autoplay, EffectFade, Navigation, Pagination]}
                   className="product-detail-swiper"
                 >
-                  {details['BlockApps-Mercata-Asset-images'].length > 0 &&
-                    details['BlockApps-Mercata-Asset-images'].map(
+                  {details['BlockApps-Asset-images'].length > 0 &&
+                    details['BlockApps-Asset-images'].map(
                       (element, index) => {
                         return (
                           <SwiperSlide>